Card 22 Jaw-Thrust Airway Translation Upload
As with card 21, name is copy/paste from DM Ukraine. The direct English translation is "removal of the jaw", maybe that should be changed, maybe it makes sense, maybe it's just a bad translation on my end & it makes sense in Ukrainian? Rest of the translations are me, through Google. I will wait a bit for this to be looked over before I merge it. Thanks.