Folder Upload Card 50 End-Tidal CO2 Monitoring
Folder upload of Card 50 End-Tidal CO2 Monitoring with colorimetric device. The Folder name is getting long, but it has to have enough of both languages to lead someone to the correct card from the folder name. Once in the folder, ENG or UKR should be enough to help them find the correct card - that was my original naming logic. Translations come from Deployed Medicine, I changed a few words, but not many. Thanks for your help.